Fixed compilation warning. JIRA CXX-8301
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@74333 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/misc/netstorage/state.cpp
@@ -1089,7 +1089,7 @@ size_t s_GetRelocateChunk(const IRegistry& registry, const string& service,
 {
     const string param_name = "relocate_chunk";
     const string service_section = s_GetSection(registry, service, param_name);
-    return registry.GetInt(service_section, param_name, default_value);
+    return registry.GetInt(service_section, param_name, static_cast<int>(default_value));
 }
